+++ src/app/components/game/game.component.ts
@@ -21,14 +21,14 @@ import { Board } from 'src/app/lib/board.library';
 export class GameComponent implements OnInit, OnChanges {
     @Output() gameDataEmitter = new EventEmitter<GameEvent>();
     @Input() game: Game;
-    @Input() gameDisplayOptions: GameDisplayOptions;
+    @Input() displayOptions: GameDisplayOptions;
 
     private _displayedMoveIndex: number;
     private boardCanvas: any;
-    private boardContext: any;
+    private _boardCtx: any;
     private boardImage: any;
     private pieceImages: any[];
-    private CURSOR_DATA: {
+    private cursor: {
         leftMouseIsDown: boolean;
         rightMouseIsDown: boolean;
         mouseOverBoard: boolean;
@@ -74,7 +74,7 @@ export class GameComponent implements OnInit, OnChanges {
     private _initPosition: Board;
 
     // event listener functions
-    private _mouseEnterEventListener: () => void;
+    private _mouseEnterEventListener: (event: any) => void;
     private _mouseLeaveEventListener: () => void;
     private _mouseMoveEventListener: (event: any) => void;
     private _mouseDownEventListener: (event: any) => void;
@@ -82,7 +82,7 @@ export class GameComponent implements OnInit, OnChanges {
 
     constructor() {
         // this.game = new Game('1k6/1p6/8/2P5/5p2/4P3/1K6/8 w - - 0 1');
-        this.CURSOR_DATA = {
+        this.cursor = {
             leftMouseIsDown: false,
             rightMouseIsDown: false,
             mouseOverBoard: false,
@@ -107,49 +107,36 @@ export class GameComponent implements OnInit, OnChanges {
         this.matchingMoves = [];
         this.displayedMoveIndex = 0;
 
-        this._mouseEnterEventListener = () => {
+        this._mouseEnterEventListener = (event: any) => {
             // just a detector of when the mouse is over the canvas object
-            this.CURSOR_DATA.mouseOverBoard = true;
+            this.cursor.mouseOverBoard = true;
+            // if mouse was down end that "drag drop event" if the mouse isn't
+            // still down
+            if (this.cursor.leftMouseIsDown) {
+                if (event.buttons === 0 || event.which !== 1) {
+                    // pause here, but this is where we put the "mouse back up"
+                    // left button
+                    this.cursor.leftMouseIsDown = false;
+                    this.cursor.dragging = false;
+                    this.cursor.draggedPieceIndex = -1;
+                    this.drawBoard();
+                    // this block is the same logic in mouseUpEvent but without
+                    // the action of doing a bunch of promotion / make move
+                    // checking
+                }
+            }
         };
         this._mouseLeaveEventListener = () => {
             // when mouse exits the canvas object
-            this.CURSOR_DATA.mouseOverBoard = false;
-            this.CURSOR_DATA.currentMousePosition = { x: -1, y: -1 };
-            this.CURSOR_DATA.overSquare = null;
+            this.cursor.mouseOverBoard = false;
+            this.cursor.currentMousePosition = { x: -1, y: -1 };
+            this.cursor.overSquare = null;
             this.tintSqFromMouseObjects = [];
-            this.CURSOR_DATA.draggedPieceIndex = -1;
+            this.cursor.draggedPieceIndex = -1;
             this.drawBoard();
         };
         this._mouseMoveEventListener = (event: any) => {
-            // condition when not on latest move
-            if (this.displayedMoveIndex !== this.game.moveHistory.length) {
-                return;
-            }
-            // this function takes the x and y coordinates of mousedata to
-            // convert that to a square coordinate
-            // we save this in an object to reference when click events happen
-            if (this.CURSOR_DATA.mouseOverBoard) {
-                this.CURSOR_DATA.currentMousePosition = this.getMousePosition(
-                    event
-                );
-                let x = this.CURSOR_DATA.currentMousePosition.x;
-                let y = this.CURSOR_DATA.currentMousePosition.y;
-                x -= x % 80;
-                y -= y % 80;
-                x /= 80;
-                y /= 80;
-                if (
-                    this.CURSOR_DATA.overSquare === null ||
-                    this.CURSOR_DATA.overSquare.x !== x ||
-                    this.CURSOR_DATA.overSquare.y !== y
-                ) {
-                    // console.log('xy', x, y);
-                    // ooh tslint taught me shorthand
-                    this.CURSOR_DATA.overSquare = { x, y };
-                    this.showMoves();
-                }
-            }
-            this.drawBoard();
+            this.mouseEventOrientationResolver(event);
         };
         this._mouseDownEventListener = (event: any) => {
             // condition when not on latest move
@@ -159,21 +146,21 @@ export class GameComponent implements OnInit, OnChanges {
             // when mouse is pressed down
             if (event.which === 1) {
                 // left button
-                this.CURSOR_DATA.leftMouseIsDown = true;
+                this.cursor.leftMouseIsDown = true;
 
-                if (this.CURSOR_DATA.overSquare) {
-                    this.CURSOR_DATA.mouseDownOn = this.CURSOR_DATA.overSquare;
+                if (this.cursor.overSquare) {
+                    this.cursor.mouseDownOn = this.cursor.overSquare;
                     if (!this.isPromoting) {
-                        this.CURSOR_DATA.dragging = true;
+                        this.cursor.dragging = true;
                     }
                     if (this.twoClickMove.attempting) {
                         // second move
-                        this.CURSOR_DATA.mouseUpOn = this.CURSOR_DATA.mouseDownOn;
-                        this.CURSOR_DATA.mouseDownOn = this.twoClickMove.source;
+                        this.cursor.mouseUpOn = this.cursor.mouseDownOn;
+                        this.cursor.mouseDownOn = this.twoClickMove.source;
                         this.attemptMoveOnBoard();
                         this.twoClickMove.attempting = false;
                         this.twoClickMove.source = null;
-                        this.CURSOR_DATA.mouseDownOn = this.CURSOR_DATA.overSquare;
+                        this.cursor.mouseDownOn = this.cursor.overSquare;
                         this.tintSqFromMouseObjects = [];
                         this.drawBoard();
                         this.showMoves();
@@ -186,7 +173,7 @@ export class GameComponent implements OnInit, OnChanges {
                 }
             } else if (event.which === 3) {
                 // right button
-                this.CURSOR_DATA.rightMouseIsDown = true;
+                this.cursor.rightMouseIsDown = true;
 
                 // logic for right mouse pressed DOWN event here
             }
@@ -199,7 +186,7 @@ export class GameComponent implements OnInit, OnChanges {
             // when mouse is released
             if (event.which === 1) {
                 // left button
-                this.CURSOR_DATA.leftMouseIsDown = false;
+                this.cursor.leftMouseIsDown = false;
 
                 // left mouse RELEASE logic here
                 if (this.isPromoting) {
@@ -207,18 +194,18 @@ export class GameComponent implements OnInit, OnChanges {
                         !this.twoClickMove.attempting &&
                         !this.twoClickMove.preventPromote
                     ) {
-                        if (this.CURSOR_DATA.overSquare) {
-                            const overSq = this.CURSOR_DATA.overSquare;
-                            this.CURSOR_DATA.mouseUpOn = overSq;
+                        if (this.cursor.overSquare) {
+                            const overSq = this.cursor.overSquare;
+                            this.cursor.mouseUpOn = overSq;
                             if (
-                                this.CURSOR_DATA.mouseDownOn.x ===
-                                    this.CURSOR_DATA.mouseUpOn.x &&
-                                this.CURSOR_DATA.mouseDownOn.y ===
-                                    this.CURSOR_DATA.mouseUpOn.y
+                                this.cursor.mouseDownOn.x ===
+                                    this.cursor.mouseUpOn.x &&
+                                this.cursor.mouseDownOn.y ===
+                                    this.cursor.mouseUpOn.y
                             ) {
                                 // console.log('', this.matchingMoves);
-                                const f = this.CURSOR_DATA.mouseDownOn.x;
-                                const r = 7 - this.CURSOR_DATA.mouseDownOn.y;
+                                const f = this.cursor.mouseDownOn.x;
+                                const r = 7 - this.cursor.mouseDownOn.y;
                                 if (f === this.matchingMoves[0].dest.file) {
                                     if (this.game.turn === Color.White) {
                                         if (r === RANK.EIGHT) {
@@ -314,28 +301,28 @@ export class GameComponent implements OnInit, OnChanges {
                         this.twoClickMove.preventPromote = false;
                     }
                 } else {
-                    if (this.CURSOR_DATA.overSquare) {
-                        this.CURSOR_DATA.mouseUpOn = this.CURSOR_DATA.overSquare;
+                    if (this.cursor.overSquare) {
+                        this.cursor.mouseUpOn = this.cursor.overSquare;
                         if (
-                            this.CURSOR_DATA.mouseDownOn.x ===
-                                this.CURSOR_DATA.mouseUpOn.x &&
-                            this.CURSOR_DATA.mouseDownOn.y ===
-                                this.CURSOR_DATA.mouseUpOn.y &&
+                            this.cursor.mouseDownOn.x ===
+                                this.cursor.mouseUpOn.x &&
+                            this.cursor.mouseDownOn.y ===
+                                this.cursor.mouseUpOn.y &&
                             this.game.getPiece(
                                 new Square(
-                                    this.CURSOR_DATA.mouseDownOn.x,
-                                    7 - this.CURSOR_DATA.mouseDownOn.y
+                                    this.cursor.mouseDownOn.x,
+                                    7 - this.cursor.mouseDownOn.y
                                 )
                             ) &&
                             this.game.getPiece(
                                 new Square(
-                                    this.CURSOR_DATA.mouseDownOn.x,
-                                    7 - this.CURSOR_DATA.mouseDownOn.y
+                                    this.cursor.mouseDownOn.x,
+                                    7 - this.cursor.mouseDownOn.y
                                 )
                             ).color === this.game.turn
                         ) {
                             this.twoClickMove.attempting = true;
-                            this.twoClickMove.source = this.CURSOR_DATA.mouseUpOn;
+                            this.twoClickMove.source = this.cursor.mouseUpOn;
                         } else {
                             this.attemptMoveOnBoard();
                         }
@@ -343,13 +330,13 @@ export class GameComponent implements OnInit, OnChanges {
                         throw new Error('mouse up not over sq');
                     }
                 }
-                this.CURSOR_DATA.dragging = false;
-                this.CURSOR_DATA.draggedPieceIndex = -1;
+                this.cursor.dragging = false;
+                this.cursor.draggedPieceIndex = -1;
                 this.drawBoard();
                 this.showMoves();
             } else if (event.which === 3) {
                 // right button
-                this.CURSOR_DATA.rightMouseIsDown = false;
+                this.cursor.rightMouseIsDown = false;
 
                 // right mouse RELEASE logic here
             }
@@ -365,7 +352,7 @@ export class GameComponent implements OnInit, OnChanges {
         this.boardCanvas.oncontextmenu = (events: any) => {
             events.preventDefault();
         };
-        this.boardContext = this.boardCanvas.getContext('2d');
+        this._boardCtx = this.boardCanvas.getContext('2d');
         this.pieceImages = [];
 
         // doing pieces first
@@ -450,7 +437,53 @@ export class GameComponent implements OnInit, OnChanges {
         );
     }
 
-    // we need to remove both of these
+    /**
+     * @description returns the mouse event, if necessary, with inverted x,y
+     * @param event - mouse event
+     * @returns any - event
+     */
+    private mouseEventOrientationResolver(event: any): any {
+        // invert x,y if private input variable indicates our orientation
+        // should be flipped
+        if (this.displayOptions.orientation === Color.Black) {
+            const offsetX = 640 - event.offsetX;
+            const offsetY = 640 - event.offsetY;
+            this.mouseMoveEvent({ offsetX, offsetY });
+        } else {
+            this.mouseMoveEvent(event);
+        }
+    }
+
+    private mouseMoveEvent(event: any) {
+        // condition when not on latest move
+        if (this.displayedMoveIndex !== this.game.moveHistory.length) {
+            return;
+        }
+        // this function takes the x and y coordinates of mousedata to
+        // convert that to a square coordinate
+        // we save this in an object to reference when click events happen
+        if (this.cursor.mouseOverBoard) {
+            this.cursor.currentMousePosition = this.getMousePosition(event);
+            let x = this.cursor.currentMousePosition.x;
+            let y = this.cursor.currentMousePosition.y;
+            x -= x % 80;
+            y -= y % 80;
+            x /= 80;
+            y /= 80;
+            if (
+                this.cursor.overSquare === null ||
+                this.cursor.overSquare.x !== x ||
+                this.cursor.overSquare.y !== y
+            ) {
+                // console.log('xy', x, y);
+                // ooh tslint taught me shorthand
+                this.cursor.overSquare = { x, y };
+                this.showMoves();
+            }
+        }
+        this.drawBoard();
+    }
+
     public setDisplayedMoveIndex(index: number): void {
         this.displayedMoveIndex = index;
     }
@@ -462,8 +495,8 @@ export class GameComponent implements OnInit, OnChanges {
     private showMoves(): void {
         const pieceMovements = this.game.getLegalMoves();
         const sq = {
-            file: this.CURSOR_DATA.overSquare.x,
-            rank: 7 - this.CURSOR_DATA.overSquare.y
+            file: this.cursor.overSquare.x,
+            rank: 7 - this.cursor.overSquare.y
         };
         this.tintSqFromMouseObjects = [];
         if (this.twoClickMove.attempting) {
@@ -483,11 +516,11 @@ export class GameComponent implements OnInit, OnChanges {
                 }
             }
         }
-        if (this.CURSOR_DATA.dragging) {
+        if (this.cursor.dragging) {
             for (const movement of pieceMovements) {
                 if (
-                    movement.src.file === this.CURSOR_DATA.mouseDownOn.x &&
-                    movement.src.rank === 7 - this.CURSOR_DATA.mouseDownOn.y
+                    movement.src.file === this.cursor.mouseDownOn.x &&
+                    movement.src.rank === 7 - this.cursor.mouseDownOn.y
                 ) {
                     this.tintSqFromMouseObjects.push({
                         dest: new Square(
@@ -503,7 +536,7 @@ export class GameComponent implements OnInit, OnChanges {
         if (
             !this.isPromoting &&
             !this.twoClickMove.attempting &&
-            !this.CURSOR_DATA.dragging
+            !this.cursor.dragging
         ) {
             for (const movement of pieceMovements) {
                 if (
@@ -547,10 +580,10 @@ export class GameComponent implements OnInit, OnChanges {
 
         for (const move of legalMoves) {
             if (
-                move.src.file === this.CURSOR_DATA.mouseDownOn.x &&
-                move.src.rank === 7 - this.CURSOR_DATA.mouseDownOn.y &&
-                move.dest.file === this.CURSOR_DATA.mouseUpOn.x &&
-                move.dest.rank === 7 - this.CURSOR_DATA.mouseUpOn.y
+                move.src.file === this.cursor.mouseDownOn.x &&
+                move.src.rank === 7 - this.cursor.mouseDownOn.y &&
+                move.dest.file === this.cursor.mouseUpOn.x &&
+                move.dest.rank === 7 - this.cursor.mouseUpOn.y
             ) {
                 this.matchingMoves.push(move);
             }
@@ -571,7 +604,7 @@ export class GameComponent implements OnInit, OnChanges {
             // console.log('invalid move attempted');
         } else {
             this.isPromoting = true;
-            if (this.CURSOR_DATA.leftMouseIsDown) {
+            if (this.cursor.leftMouseIsDown) {
                 this.twoClickMove.preventPromote = true;
             } else {
                 this.twoClickMove.preventPromote = false;
@@ -579,19 +612,9 @@ export class GameComponent implements OnInit, OnChanges {
         }
     }
 
-    private getMousePosition(events: any): { x: number; y: number } {
-        let obj = this.boardCanvas;
-        let top = 0;
-        let left = 0;
-        let mX = 0;
-        let mY = 0;
-        while (obj && obj.tagName !== 'BODY') {
-            top += obj.offsetTop;
-            left += obj.offsetLeft;
-            obj = obj.offsetParent;
-        }
-        mX = events.clientX - left + window.pageXOffset;
-        mY = events.clientY - top + window.pageYOffset;
+    private getMousePosition(event: any): { x: number; y: number } {
+        const mX = event.offsetX;
+        const mY = event.offsetY;
         return { x: mX, y: mY };
     }
 
@@ -599,17 +622,18 @@ export class GameComponent implements OnInit, OnChanges {
      * @description triggers a redraw event of the entire board
      */
     public drawBoard(): void {
-        this.boardContext.restore();
-        this.boardContext.globalAlpha = 1;
-        // this.boardContext.drawImage(this.boardImage, 0, 0);
-        this.boardContext.font = '15px Arial';
+        this.boardCtx.restore();
+        this.boardCtx.globalAlpha = 1;
+        // this.boardCtx.drawImage(this.boardImage, 0, 0);
+        this.boardCtx.font = '15px Arial';
         for (let i = 0; i < 8; i++) {
             for (let j = 0; j < 8; j++) {
-                this.boardContext.fillStyle =
+                // no need to invert this if displayOptions.orientation is black
+                this.boardCtx.fillStyle =
                     (i + j) % 2 === 0
-                        ? this.gameDisplayOptions.colorScheme.light
-                        : this.gameDisplayOptions.colorScheme.dark;
-                this.boardContext.fillRect(i * 80, j * 80, 80, 80);
+                        ? this.displayOptions.colorScheme.light
+                        : this.displayOptions.colorScheme.dark;
+                this.boardCtx.fillRect(i * 80, j * 80, 80, 80);
             }
         }
         for (let i = 0; i < 8; i++) {
@@ -617,62 +641,94 @@ export class GameComponent implements OnInit, OnChanges {
                 this.refreshCanvasSquare(i, j);
             }
         }
-        if (this.gameDisplayOptions.showCoordinates) {
-            for (let i = 0; i < 8; i++) {
-                for (let j = 0; j < 8; j++) {
-                    if ((i + j) % 2 === 0) {
-                        // light sq
-                        this.boardContext.fillStyle = this.gameDisplayOptions.colorScheme.dark;
-                    } else {
-                        // dark sq
-                        this.boardContext.fillStyle = this.gameDisplayOptions.colorScheme.light;
-                    }
-                    if (i === 7) {
-                        // right side
-                        this.boardContext.fillText(
-                            '' + (8 - j),
-                            628,
-                            j * 80 + 15
-                        );
-                    }
-                    if (j === 7) {
-                        // bottom row
-                        this.boardContext.fillText(
-                            fileToString(i),
-                            i * 80 + 5,
-                            635
-                        );
-                    }
-                }
-            }
+        if (this.displayOptions.showCoordinates) {
+            this.drawCoordinates();
         }
-        this.boardContext.globalAlpha = 1;
-        if (this.CURSOR_DATA.draggedPieceIndex !== -1) {
-            this.boardContext.drawImage(
-                this.pieceImages[this.CURSOR_DATA.draggedPieceIndex],
-                this.CURSOR_DATA.currentMousePosition.x - 40,
-                this.CURSOR_DATA.currentMousePosition.y - 40
+        this.boardCtx.globalAlpha = 1;
+        if (
+            this.cursor.draggedPieceIndex !== -1 &&
+            this.cursor.mouseOverBoard
+        ) {
+            let x = this.cursor.currentMousePosition.x;
+            let y = this.cursor.currentMousePosition.y;
+            if (this.displayOptions.orientation === Color.Black) {
+                x = 640 - 1 - x;
+                y = 640 - 1 - y;
+            }
+            this.boardCtx.drawImage(
+                this.pieceImages[this.cursor.draggedPieceIndex],
+                x - 40,
+                y - 40
             );
         }
         if (this.isPromoting) {
-            this.boardContext.globalAlpha = 0.5;
-            this.boardContext.fillStyle = 'black';
-            this.boardContext.fillRect(0, 0, 640, 640);
-            this.boardContext.globalAlpha = 1;
-            const x = this.matchingMoves[0].dest.file;
-            this.boardContext.fillStyle = '#AAAAAA';
+            this.boardCtx.globalAlpha = 0.5;
+            this.boardCtx.fillStyle = 'black';
+            this.boardCtx.fillRect(0, 0, 640, 640);
+            this.boardCtx.globalAlpha = 1;
+            // this is where we do the invert trickery
+            let drawAtTopOfBoard = this.game.turn === Color.White;
+            let x = this.matchingMoves[0].dest.file;
+            if (this.displayOptions.orientation === Color.Black) {
+                // flip
+                drawAtTopOfBoard = !drawAtTopOfBoard;
+                x = 7 - x;
+            }
+            this.boardCtx.fillStyle = '#AAAAAA';
+            this.boardCtx.fillRect(
+                x * 80,
+                drawAtTopOfBoard ? 0 : 80 * 4,
+                80,
+                80 * 4
+            );
+            const _4ys = drawAtTopOfBoard
+                ? [0, 80, 80 * 2, 80 * 3]
+                : [80 * 7, 80 * 6, 80 * 5, 80 * 4];
             if (this.game.turn === Color.White) {
-                this.boardContext.fillRect(x * 80, 0, 80, 320);
-                this.boardContext.drawImage(this.pieceImages[1], x * 80, 0);
-                this.boardContext.drawImage(this.pieceImages[3], x * 80, 80);
-                this.boardContext.drawImage(this.pieceImages[4], x * 80, 160);
-                this.boardContext.drawImage(this.pieceImages[2], x * 80, 240);
+                this.boardCtx.drawImage(this.pieceImages[1], x * 80, _4ys[0]);
+                this.boardCtx.drawImage(this.pieceImages[3], x * 80, _4ys[1]);
+                this.boardCtx.drawImage(this.pieceImages[4], x * 80, _4ys[2]);
+                this.boardCtx.drawImage(this.pieceImages[2], x * 80, _4ys[3]);
             } else {
-                this.boardContext.fillRect(x * 80, 320, 80, 320);
-                this.boardContext.drawImage(this.pieceImages[7], x * 80, 560);
-                this.boardContext.drawImage(this.pieceImages[9], x * 80, 480);
-                this.boardContext.drawImage(this.pieceImages[10], x * 80, 400);
-                this.boardContext.drawImage(this.pieceImages[8], x * 80, 320);
+                this.boardCtx.drawImage(this.pieceImages[7], x * 80, _4ys[0]);
+                this.boardCtx.drawImage(this.pieceImages[9], x * 80, _4ys[1]);
+                this.boardCtx.drawImage(this.pieceImages[10], x * 80, _4ys[2]);
+                this.boardCtx.drawImage(this.pieceImages[8], x * 80, _4ys[3]);
+            }
+        }
+    }
+
+    private drawCoordinates(): void {
+        for (let i = 0; i < 8; i++) {
+            for (let j = 0; j < 8; j++) {
+                // light/dark sq
+                this.boardCtx.fillStyle =
+                    (i + j) % 2 === 0
+                        ? this.displayOptions.colorScheme.dark
+                        : this.displayOptions.colorScheme.light;
+                if (i === 7) {
+                    // right side
+                    this.boardCtx.fillText(
+                        '' +
+                            (this.displayOptions.orientation === Color.White
+                                ? 8 - j
+                                : j + 1),
+                        628,
+                        j * 80 + 15
+                    );
+                }
+                if (j === 7) {
+                    // bottom row
+                    this.boardCtx.fillText(
+                        fileToString(
+                            this.displayOptions.orientation === Color.White
+                                ? i
+                                : 7 - i
+                        ),
+                        i * 80 + 5,
+                        635
+                    );
+                }
             }
         }
     }
@@ -685,7 +741,7 @@ export class GameComponent implements OnInit, OnChanges {
                       this.displayedMoveIndex - 1
                   ].resultingBoard.getPiece(new Square(x, y));
         if (this.displayedMoveIndex !== 0) {
-            let lastMove = this.game.moveHistory[this.displayedMoveIndex - 1];
+            const lastMove = this.game.moveHistory[this.displayedMoveIndex - 1];
             if (
                 (lastMove.src.file === x && lastMove.src.rank === y) ||
                 (lastMove.dest.file === x && lastMove.dest.rank === y)
@@ -694,18 +750,18 @@ export class GameComponent implements OnInit, OnChanges {
             }
         }
         if (
-            this.CURSOR_DATA.overSquare &&
-            this.CURSOR_DATA.overSquare.x === x &&
-            this.CURSOR_DATA.overSquare.y === 7 - y &&
+            this.cursor.overSquare &&
+            this.cursor.overSquare.x === x &&
+            this.cursor.overSquare.y === 7 - y &&
             !this.isPromoting &&
             !(
                 this.twoClickMove.source &&
-                this.CURSOR_DATA.overSquare.x === this.twoClickMove.source.x &&
-                this.CURSOR_DATA.overSquare.y === this.twoClickMove.source.y
+                this.cursor.overSquare.x === this.twoClickMove.source.x &&
+                this.cursor.overSquare.y === this.twoClickMove.source.y
             )
         ) {
             this.tintSquare(x, 7 - y, 'gray', 0.5);
-            this.boardContext.globalAlpha = 1; // reset this to full
+            this.boardCtx.globalAlpha = 1; // reset this to full
         }
         if (
             this.twoClickMove.attempting &&
@@ -718,15 +774,23 @@ export class GameComponent implements OnInit, OnChanges {
                 'yellow',
                 0.5
             );
-            this.boardContext.globalAlpha = 1; // reset this to full
+            this.boardCtx.globalAlpha = 1; // reset this to full
+
+            // TODO currently LEAVING this if block alone, but I don't think it
+            // actually does anything...we already draw the pieces, investigate
+            // commenting this whole block out.
             if (piece) {
                 const color = piece.color;
                 const pieceType = piece.type;
                 const index = (color ? 6 : 0) + pieceType;
-                this.boardContext.drawImage(
+                this.boardCtx.drawImage(
                     this.pieceImages[index],
-                    x * 80,
-                    (7 - y) * 80
+                    (this.displayOptions.orientation === Color.White
+                        ? x
+                        : 7 - x) * 80,
+                    (this.displayOptions.orientation === Color.White
+                        ? 7 - y
+                        : y) * 80
                 );
             }
         }
@@ -738,22 +802,26 @@ export class GameComponent implements OnInit, OnChanges {
                 tintSq.gA
             );
         }
-        this.boardContext.globalAlpha = 1;
+        this.boardCtx.globalAlpha = 1;
         if (piece) {
             const color = piece.color;
             const pieceType = piece.type;
             const index = (color === 1 ? 6 : 0) + pieceType;
             if (
-                this.CURSOR_DATA.dragging &&
-                this.CURSOR_DATA.mouseDownOn.x === x &&
-                this.CURSOR_DATA.mouseDownOn.y === 7 - y
+                this.cursor.dragging &&
+                this.cursor.mouseDownOn.x === x &&
+                this.cursor.mouseDownOn.y === 7 - y
             ) {
-                this.CURSOR_DATA.draggedPieceIndex = index;
+                this.cursor.draggedPieceIndex = index;
             } else {
-                this.boardContext.drawImage(
+                this.boardCtx.drawImage(
                     this.pieceImages[index],
-                    x * 80,
-                    (7 - y) * 80
+                    (this.displayOptions.orientation === Color.White
+                        ? x
+                        : 7 - x) * 80,
+                    (this.displayOptions.orientation === Color.White
+                        ? 7 - y
+                        : y) * 80
                 );
             }
         }
@@ -765,9 +833,13 @@ export class GameComponent implements OnInit, OnChanges {
         color: string,
         globalAlpha: number
     ): void {
-        this.boardContext.globalAlpha = globalAlpha;
-        this.boardContext.fillStyle = color;
-        this.boardContext.fillRect(x * 80, y * 80, 80, 80);
+        this.boardCtx.globalAlpha = globalAlpha;
+        this.boardCtx.fillStyle = color;
+        if (this.displayOptions.orientation === Color.Black) {
+            x = 7 - x;
+            y = 7 - y;
+        }
+        this.boardCtx.fillRect(x * 80, y * 80, 80, 80);
     }
 
     public getGame(): Game {
@@ -795,9 +867,9 @@ export class GameComponent implements OnInit, OnChanges {
         if (times < 1) {
             return;
         }
-        let redTint = {
+        const redTint = {
             dest: new Square(sq.file, 7 - sq.rank),
-            color: color,
+            color,
             gA: 0.5
         };
         this.tintSqData.push(redTint);
@@ -816,5 +888,8 @@ export class GameComponent implements OnInit, OnChanges {
     }
     set displayedMoveIndex(i: number) {
         this._displayedMoveIndex = i;
+    }
+    get boardCtx(): any {
+        return this._boardCtx;
     }
 }