+++ src/app/components/opening-training-game/opening-training-game.component.ts
@@ -8,6 +8,7 @@ import {
 } from 'src/app/lib/interface.library';
 import { Openings } from 'src/app/data/openings';
 import { GameComponent } from '../game/game.component';
+import { Color } from 'src/app/lib/util.library';
 
 @Component({
     selector: 'app-opening-training-game',
@@ -38,6 +39,7 @@ export class OpeningTrainingGameComponent implements OnInit {
     constructor() {
         this.game = new Game();
         this._gameDisplayOptions = {
+            orientation: Color.White,
             showCoordinates: true,
             colorScheme: {
                 light: '#f0d9b9',